+++ src/events/ready.test.ts
@@ -30,26 +30,11 @@ import type { ClientPresence } from 'discord.js';
 import { Client } from 'discord.js';
 const client = new Client({ intents: [] });
 
-// Mock parseArgs so we can control what the args are
-import type { parseArgs } from '../helpers/parseArgs.js';
-const mockParseArgs = vi.hoisted(() => vi.fn<typeof parseArgs>());
-vi.mock('../helpers/parseArgs', () => ({ parseArgs: mockParseArgs }));
-
-// Mock deployCommands so we can track it
-vi.mock('../helpers/actions/deployCommands.js');
-import { deployCommands } from '../helpers/actions/deployCommands.js';
-const mockDeployCommands = deployCommands as Mock<typeof deployCommands>;
-
-// Mock revokeCommands so we can track it
-vi.mock('../helpers/actions/revokeCommands.js');
-import { revokeCommands } from '../helpers/actions/revokeCommands.js';
-const mockRevokeCommands = revokeCommands as Mock<typeof revokeCommands>;
-
-// Mock verifyCommandDeployments so we can track it
-vi.mock('../helpers/actions/verifyCommandDeployments.js');
-import { verifyCommandDeployments } from '../helpers/actions/verifyCommandDeployments.js';
-const mockVerifyCommandDeployments = verifyCommandDeployments as Mock<
-	typeof verifyCommandDeployments
+// Mock ensureCommandDeployments so we can track it
+vi.mock('../helpers/actions/ensureCommandDeployments.js');
+import { ensureCommandDeployments } from '../helpers/actions/ensureCommandDeployments.js';
+const mockEnsureCommandDeployments = ensureCommandDeployments as Mock<
+	typeof ensureCommandDeployments
 >;
 
 // Mock the logger so nothing is printed
@@ -60,11 +45,6 @@ import { ready } from './ready.js';
 
 describe('once(ready)', () => {
 	beforeEach(() => {
-		// Default is no deploy, no revoke, no method behavior
-		mockParseArgs.mockReturnValue({
-			deploy: false,
-			revoke: false,
-		});
 		mockSetActivity.mockReturnValue({} as ClientPresence);
 	});
 
@@ -72,49 +52,9 @@ describe('once(ready)', () => {
 		vi.restoreAllMocks();
 	});
 
-	test("doesn't touch commands if the `deploy` and `revoke` flags are not set", async () => {
-		mockParseArgs.mockReturnValue({
-			deploy: false,
-			revoke: false,
-		});
+	test('runs command deployments', async () => {
 		await ready.execute(client);
-		expect(mockDeployCommands).not.toHaveBeenCalled();
-		expect(mockRevokeCommands).not.toHaveBeenCalled();
-	});
-
-	test('deploys commands if the `deploy` flag is set', async () => {
-		mockParseArgs.mockReturnValue({
-			deploy: true,
-			revoke: false,
-		});
-		await ready.execute(client);
-		expect(mockDeployCommands).toHaveBeenCalledWith(client);
-		expect(mockRevokeCommands).not.toHaveBeenCalled();
-	});
-
-	test('revokes commands if the `revoke` flag is set', async () => {
-		mockParseArgs.mockReturnValue({
-			deploy: false,
-			revoke: true,
-		});
-		await ready.execute(client);
-		expect(mockDeployCommands).not.toHaveBeenCalled();
-		expect(mockRevokeCommands).toHaveBeenCalledWith(client);
-	});
-
-	test('deploys commands if both the `revoke` and `deploy` flags are set', async () => {
-		mockParseArgs.mockReturnValue({
-			deploy: true,
-			revoke: true,
-		});
-		await ready.execute(client);
-		expect(mockDeployCommands).toHaveBeenCalledWith(client);
-		expect(mockRevokeCommands).not.toHaveBeenCalled();
-	});
-
-	test('verifies command deployments', async () => {
-		await ready.execute(client);
-		expect(mockVerifyCommandDeployments).toHaveBeenCalledWith(client);
+		expect(mockEnsureCommandDeployments).toHaveBeenCalledWith(client);
 	});
 
 	test('sets user activity', async () => {